// Aida_Angular/src/app/shared/models/employee-data.ts
export interface EmployeeModel
{
    code: string ;
    firstName: string ;
    lastName: string ;
    nationality: string ;
    birthDate: string ;
    gender: string ;
    position: string ;
    cardType: string ;
    cardNumber: string ;
    address1: string ;
    address2: string ;
    city: string ;
    country: string ;
    pincode: string ;
    email: string ;
    mobile: string ;
    phone: string ;
    loginRole: string ;
    userpassword: string ;
}